+++ quimb/gates.py
@@ -23,3 +23,5 @@ CZ = operators.cZ()
 
 SWAP = operators.swap()
 ISWAP = operators.iswap()
+
+U3 = operators.U_gate